backend/requirements.txt
fastapi
uvicorn
sqlalchemy
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
pydantic-settings
python-dotenv
pyotp
qrcode
email-validator
fastapi-mail
